ZERO-74 [Rename]: 폴더 구조에 맞춰 파일 이동
